+++ src/core/jslib.c
@@ -279,7 +279,10 @@ EM_JS_VAL(JsVal, JsvObject_Values, (JsVal obj), {
 
 EM_JS_VAL(JsVal,
 JsvObject_toString, (JsVal obj), {
-  return obj.toString();
+  if (hasMethod(obj, "toString")) {
+    return obj.toString();
+  }
+  return Object.prototype.toString.call(obj);
 });